(GH-11) Add markdownlint & reflow extensions
This change adds the markdownlint and reflow-markdown extensions to the
list of recommendations for the project workspace. The markdownlint
extension can be used to identify and automatically fix many rule
violations.

The reflow-markdown extension can be used to more easily reflow
Markdown to a specified line length. In this change, the line length is
set to 100 in the workspace settings to make it easier to see and manage
reflowing text.

Reflowed markdown is easier to diff during review. Both git and the
GitHub UI will sometimes fail if a single line diff is too long or
complicated. Reflowed Markdown largely obviates this problem.

This change also adds code rulers at columns 76, 80, 96, and 100. This
provides a visual cue when line length is approaching/past the maximum
for code blocks and prose, respectively. While markdownlint will also
warn in the editor when a line is too long, it's often useful to see
this before encountering it while authoring longer passages.