(GH-11) Migrate to reference links
Prior to this change, all of the links in the Markdown documents were
inline links. This could make parsing paragraphs difficult and meant in
some cases that links were repeated, requiring them to be updated in
multiple places if the URLs changed in the future.

This change migrates the markdown link syntax to reference link syntax,
placing the reference link definitions at the bottom of each file. It
uses a two-digit order-of-appearance numbering, where links are defined
as `[foo][01]` and `[bar][02]`. This limits the space used by the link
references while also allowing reuse.

A possible alternative would be to leave the image links as inline, but
this change standardizes the links across every file and link type.

While this change was performed manually, it should be possible in the
future to use a scripted action or the continually improving options to
[extract link definitions][01] and [organize link definitions][02] in
the Markdown support for VS Code.

[01]: https://code.visualstudio.com/updates/v1_72#_extract-to-link-definition-for-markdown
[02]: https://code.visualstudio.com/updates/v1_72#_organize-link-definitions-in-markdown